adjust int max range

+++ parSumRows.c
@@ -2,6 +2,7 @@
 #include <stdlib.h>
 #include <time.h>
 #include <string.h>
+#include <stdint.h>  // Include this header for uint64_t
 #include <mpi.h>
 #include <omp.h>
 
@@ -11,13 +12,13 @@
 // Function to generate a random matrix of integers
 void generate_random_matrix(int n, int *matrix) {
     for (int i = 0; i < n * n; i++) {
-        matrix[i] = rand() % MAX_RAND_RANGE; 
+        matrix[i] = rand() % MAX_RAND_RANGE;
     }
 }
 
 // Function to sum the elements of a row
-int sum_row(int n, int *row) {
-    int sum = 0;
+uint64_t sum_row(int n, int *row) {
+    uint64_t sum = 0;  // Change data type to uint64_t
     for (int i = 0; i < n; i++) {
         sum += row[i];
     }
@@ -27,7 +28,7 @@ int sum_row(int n, int *row) {
 int main(int argc, char *argv[]) {
 
     double par_start_time, par_end_time;
-    
+
     if (argc != 2) {
         fprintf(stderr, "Usage: %s <matrix_size>\n", argv[0]);
         return 1;
@@ -39,7 +40,6 @@ int main(int argc, char *argv[]) {
         fprintf(stderr, "Invalid matrix size\n");
         return 1;
     }
-    
 
     MPI_Init(&argc, &argv);
 
@@ -48,7 +48,7 @@ int main(int argc, char *argv[]) {
     MPI_Comm_size(MPI_COMM_WORLD, &size);
 
     // Set the number of threads for parallelization
-    omp_set_num_threads(OMP_NUM_THREADS); 
+    omp_set_num_threads(OMP_NUM_THREADS);
 
     int *matrix = NULL;
 
@@ -71,31 +71,30 @@ int main(int argc, char *argv[]) {
 
     // Allocate memory for the local rows on each processor
     int *local_rows = (int *)malloc(rows_per_proc * n * sizeof(int));
-    
+
     // Distribute rows among processors
     MPI_Scatter(matrix, rows_per_proc * n, MPI_INT, local_rows, rows_per_proc * n, MPI_INT, 0, MPI_COMM_WORLD);
 
     // Perform row sums in parallel
-    #pragma omp parallel for
+#pragma omp parallel for
     for (int i = 0; i < rows_per_proc; i++) {
-        int row_sum = sum_row(n, &local_rows[i * n]);
+        uint64_t row_sum = sum_row(n, &local_rows[i * n]);  // Change data type to uint64_t
 
         char processor_name[MPI_MAX_PROCESSOR_NAME];
         int name_len;
         MPI_Get_processor_name(processor_name, &name_len);
 
         // Print the sum of each row
-        printf("Processor %d is running on host %s - Row %d Sum: %d\n", rank, processor_name, i + rank * rows_per_proc, row_sum);
+        printf("Processor %d is running on host %s - Row %d Sum: %llu\n", rank, processor_name, i + rank * rows_per_proc, row_sum);
 
     }
 
     free(matrix);
     free(local_rows);
 
-
     MPI_Finalize();
 
-    if(rank == 0){
+    if (rank == 0) {
         par_end_time = omp_get_wtime();
         printf("Parallel time: %lfs\n", par_end_time - par_start_time);
         printf("\n--------------------------------\n");
